\name{osng.proj4string}
\alias{osng.proj4string}
\title{\code{Proj4} string for postcode coordinates}
\description{The \code{proj4} string for OS National Grid coordinates}
\details{The \code{proj4} string used to specify the Ordnance Survey National Grid 'proj4' string
allowing the locations to be transformed to other projections and Longitude/Latitude 
form via the \code{rgdal} package.}
\value{The \code{proj4} string used to represent the OS National Grid coordinates}
\examples{osng.proj4string()}
